requirements.txt
# ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
# FestivalFlow AI — requirements.txt
# Python 3.11対応 / Streamlit Cloud デプロイ確認済み
# ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━

# ── コアフレームワーク ──────────────────────────
streamlit==1.35.0           # メインWebフレームワーク。st.session_stateで状態管理

# ── データ処理 ──────────────────────────────────
pandas==2.2.2               # イベントデータのDataFrame管理・CSV出力処理
numpy==1.26.4               # M/M/1数式の数値計算・モンテカルロシミュレーション

# ── 可視化 ──────────────────────────────────────
plotly==5.22.0              # インタラクティブなヒートマップ・時系列グラフ
altair==5.3.0               # 宣言的グラフ記法。KPIダッシュボードの補助チャート

# ── 数理最適化・統計 ────────────────────────────
scipy==1.13.1               # Erlang-C公式・統計分布。M/M/cモデル拡張時に使用

# ── セキュリティ ────────────────────────────────
bcrypt==4.1.3               # PINのbcryptハッシュ化（SHA-256より安全なソルト付きハッシュ）
python-jose==3.3.0          # JWTトークン生成・検証（セッション管理）

# ── 環境変数管理 ────────────────────────────────
python-dotenv==1.0.1        # .envファイル読み込み（ローカル開発用）

# ── バックエンド（オプション・アダプター設計） ───
supabase==2.4.6             # SupabaseクライアントSDK（未接続時はモックデータで動作）

# ── 開発ツール ──────────────────────────────────
pytest==8.2.2               # 自動テストランナー
pytest-cov==5.0.0           # カバレッジレポート（ポートフォリオ用：80%以上を目標）
black==24.4.2               # コードフォーマッター（PEP8準拠）
ruff==0.4.7                 # 高速リンター（flake8/isort統合代替）
